refactor cu_full
