+++ templates/cu_full/.gitignore
@@ -1,0 +1,3 @@
+/target
+*.copper
+Cargo.lock
