+++ templates/cu_full/Cargo.toml.template
@@ -1,11 +1,7 @@
 [workspace]
 members = [
-    "apps/{{project-name}}",
-    "apps/{{project-name}}-demo",
-    "components/payloads/{{project-name}}-types",
-    "components/sources/{{project-name}}-source",
-    "components/tasks/{{project-name}}-processor",
-    "components/sinks/{{project-name}}-sink",
+    "apps/cu_example_app",
+    "components/bridges/cu_example_shared_bridge",
 ]
 resolver = "2"
