+++ templates/cu_full/apps/README.md
@@ -1,0 +1,1 @@
+Application crates live here. `cu_example_app` is the starter app and owns its runtime config and logs.
