+++ templates/cu_full/apps/cu_example_app/README.md
@@ -1,0 +1,1 @@
+This is the primary Copper application for the workspace (`cu_example_app`). Local tasks and messages live under `src/`.
