+++ templates/cu_full/apps/cu_example_app/build.rs
