+++ templates/cu_full/apps/cu_example_app/src/messages.rs
@@ -1,0 +1,8 @@
+use bincode::{Decode, Encode};
+use cu29::prelude::*;
+
+// Define a message type
+#[derive(Default, Debug, Clone, Encode, Decode, Serialize)]
+pub struct MyPayload {
+    pub value: i32,
+}
