+++ templates/cu_full/apps/cu_example_app/src/tasks/mod.rs
@@ -1,0 +1,7 @@
+mod local_example_sink;
+mod local_example_src;
+mod local_example_task;
+
+pub use local_example_sink::MySink;
+pub use local_example_src::MySource;
+pub use local_example_task::MyTask;
