--- templates/cu_full/apps/{{project-name}}-demo/Cargo.toml.template
@@ -1,30 +1,0 @@
-[package]
-name = "{{project-name|kebab_case}}-demo"
-version = "0.1.0"
-edition = "2021"
-default-run = "{{project-name|kebab_case}}-demo"
-
-# The main executable of your application
-[[bin]]
-name = "{{project-name|kebab_case}}-demo"
-path = "src/main.rs"
-
-# A custom made log reader application for your application.
-[[bin]]
-name = "{{project-name|kebab_case}}-demo-logreader"
-path = "src/logreader.rs"
-required-features = ["logreader"]
-
-[features]
-default = []
-logreader = ["dep:cu29-export"]
-
-[dependencies]
-cu29 = { workspace = true }
-cu29-helpers = { workspace = true }
-cu29-export = { workspace = true, optional = true }
-serde = { workspace = true }
-{{project-name|kebab_case}}-types = { path = "../../components/payloads/{{project-name}}-types" }
-{{project-name|kebab_case}}-source = { path = "../../components/sources/{{project-name}}-source" }
-{{project-name|kebab_case}}-processor = { path = "../../components/tasks/{{project-name}}-processor" }
-{{project-name|kebab_case}}-sink = { path = "../../components/sinks/{{project-name}}-sink" }
