--- templates/cu_full/apps/{{project-name}}-demo/bin/README.md
@@ -1,1 +1,0 @@
-Helpers moved to `justfile` at the workspace root.
