--- templates/cu_full/apps/{{project-name}}-demo/config/copperconfig.ron
@@ -1,27 +1,0 @@
-// A sample configuration to get you started.
-(
-    tasks: [
-        (
-            id: "src",
-            type: "{{project-name|snake_case}}_source::MySource",
-        ),
-        (
-            id: "t-1",
-            type: "{{project-name|snake_case}}_processor::MyTask",
-        ),
-        (
-            id: "t-2",
-            type: "{{project-name|snake_case}}_processor::MyTask",
-        ),
-        (
-            id: "sink",
-            type: "{{project-name|snake_case}}_sink::MySink",
-        ),
-     ],
-    // Builds a small pipeline so you can see task reuse across apps.
-    cnx: [
-        (src: "src",  dst: "t-1",  msg: "{{project-name|snake_case}}_types::MyPayload"),
-        (src: "t-1",  dst: "t-2",  msg: "{{project-name|snake_case}}_types::MyPayload"),
-        (src: "t-2",  dst: "sink", msg: "{{project-name|snake_case}}_types::MyPayload"),
-    ],
-)
