--- templates/cu_full/apps/{{project-name}}-demo/src/logreader.rs
@@ -1,11 +1,0 @@
-use cu29::prelude::*;
-use cu29_export::run_cli;
-
-// This will create the CuStampedDataSet that is specific to your copper project.
-// It is used to instruct the log reader how to decode the logs.
-gen_cumsgs!("config/copperconfig.ron");
-
-#[cfg(feature = "logreader")]
-fn main() {
-    run_cli::<CuStampedDataSet>().expect("Failed to run the export CLI");
-}
