--- templates/cu_full/apps/{{project-name}}/bin/README.md
@@ -1,1 +1,0 @@
-Helpers moved to `justfile` at the workspace root.
