+++ templates/cu_full/components/README.md
@@ -1,0 +1,1 @@
+Shared components live here (bridges, sources, tasks, payloads, and more). The `cu_example_shared_bridge` crate is a starter bridge.
