+++ templates/cu_full/components/bridges/cu_example_shared_bridge/README.md
@@ -1,0 +1,3 @@
+# cu_example_shared_bridge
+
+Shared bridge crate for the workspace. Use this as a starting point for bridge logic and shared message types.
