+++ templates/cu_full/components/bridges/cu_example_shared_bridge/src/lib.rs
@@ -1,0 +1,2 @@
+pub mod messages;
+pub use messages::SharedBridgePayload;
