+++ templates/cu_full/components/bridges/cu_example_shared_bridge/src/messages.rs
@@ -1,0 +1,8 @@
+use bincode::{Decode, Encode};
+use cu29::prelude::*;
+
+// Shared bridge message type
+#[derive(Default, Debug, Clone, Encode, Decode, Serialize)]
+pub struct SharedBridgePayload {
+    pub value: i32,
+}
