--- templates/cu_full/components/payloads/{{project-name}}-types/src/lib.rs
@@ -1,8 +1,0 @@
-use bincode::{Decode, Encode};
-use serde::Serialize;
-
-// Define shared message types here.
-#[derive(Default, Debug, Clone, Encode, Decode, Serialize)]
-pub struct MyPayload {
-    pub value: i32,
-}
