--- templates/cu_full/components/sinks/{{project-name}}-sink/build.rs
@@ -1,6 +1,0 @@
-fn main() {
-    println!(
-        "cargo:rustc-env=LOG_INDEX_DIR={}",
-        std::env::var("OUT_DIR").unwrap()
-    );
-}
