--- templates/cu_full/components/sources/{{project-name}}-source/Cargo.toml.template
@@ -1,8 +1,0 @@
-[package]
-name = "{{project-name|kebab_case}}-source"
-version = "0.1.0"
-edition = "2021"
-
-[dependencies]
-cu29 = { workspace = true }
-{{project-name|kebab_case}}-types = { path = "../../payloads/{{project-name}}-types" }
