--- templates/cu_full/docs/README.md
@@ -1,1 +1,0 @@
-Project documentation, design notes, and decision records live here.
