+++ templates/cu_full/justfile
@@ -1,25 +1,8 @@
 # Workspace automation helpers.
 
-# Extract the structured log via the log reader.
-log:
-  #!/usr/bin/env bash
-  set -euo pipefail
-  APP_NAME="${APP_NAME:-{{project-name|kebab_case}}}"
-  APP_DIR="${APP_DIR:-{{project-name}}}"
-  RUST_BACKTRACE=1 cargo run --features=logreader -p "$APP_NAME" --bin "${APP_NAME}-logreader" logs/"${APP_NAME}".copper extract-text-log target/debug/cu29_log_index
-
-# Extract CopperLists from the log output.
-cl:
-  #!/usr/bin/env bash
-  set -euo pipefail
-  APP_NAME="${APP_NAME:-{{project-name|kebab_case}}}"
-  APP_DIR="${APP_DIR:-{{project-name}}}"
-  RUST_BACKTRACE=1 cargo run --features=logreader -p "$APP_NAME" --bin "${APP_NAME}-logreader" logs/"${APP_NAME}".copper extract-copperlists
-
 # Render the execution DAG from the app config.
 rcfg:
   #!/usr/bin/env bash
   set -euo pipefail
-  APP_NAME="${APP_NAME:-{{project-name|kebab_case}}}"
-  APP_DIR="${APP_DIR:-{{project-name}}}"
-  {{copper_root_path}}/target/debug/cu29-rendercfg apps/"${APP_DIR}"/config/copperconfig.ron --open
+  APP_DIR="${APP_DIR:-cu_example_app}"
+  {{copper_root_path}}/target/debug/cu29-rendercfg apps/"${APP_DIR}"/copperconfig.ron --open
