+++ templates/cu_project/.gitignore
@@ -1,0 +1,3 @@
+/target
+*.copper
+Cargo.lock
